Refactor cli-client code
